flaskAPI/requirements.txt
flask
flask-mysqldb
flask-cors
ultralytics
opencv-python
pillow